requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
flake8==3.8.4
idna==2.10
mccabe==0.6.1
multidict==5.1.0
pip==20.3.3
pip-autoremove==0.9.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
setuptools==52.0.0.post20210125
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wheel==0.36.2
yarl==1.6.3
